tests/wpt/mozilla/tests/bluetooth/requestDevice/adapter-off.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.not_powered);
    return promise_rejects_dom(
        t, 'NotFoundError', window.navigator.bluetooth.requestDevice({filters: [{services: [generic_access.name]}]}),
        'Bluetooth adapter is not powered.');
}, 'Reject with NotFoundError if the adapter is off.');
</script>
